Allow to change the maximum stack size of custom vanilla item mappings
